# gomino-src/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorResponse:
    properties:
      api:message:
        type: string
      api:statuscode:
        type: integer
    type: object
...
  user.CreateCommentRequest:
    properties:
      content:
        type: string
      respondTo:
        type: string
      stickerId:
        type: string
      type:
        type: integer
    type: object
...
  user.UpdateUserRequest:
    properties:
      avatarFrameId:
        type: string
      content:
        type: string
      icon:
        type: string
      mediaList:
        type: string
      mood:
        type: string
      moodSticker:
        type: string
      nickname:
        type: string
      notificationSubscriptionStatus:
        type: integer
      pushEnabled:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@gomino.com
    name: API Support
  description: This is the API server for GoMino.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoMino API
  version: "1.0"
paths:
...
  /g/s/user-profile/{userId}:
    get:
      consumes:
      - application/json
      description: Get public user profile information
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user profile
      tags:
      - user-profile
    post:
      consumes:
      - application/json
      description: Update user profile details
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update user profile
      tags:
      - user-profile
  /g/s/user-profile/{userId}/comment:
    post:
      consumes:
      - application/json
      description: Reply to a comment on the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Reply to wall comment
      tags:
      - user-profile
  /g/s/user-profile/{userId}/comment/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment from the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete wall comment
      tags:
      - user-profile
  /g/s/user-profile/{userId}/g-comment:
    get:
      consumes:
      - application/json
      description: Get comments on the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Sort order (newest/oldest)
        in: query
        name: sort
        type: string
      - description: Start offset
        in: query
        name: start
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get wall comments
      tags:
      - user-profile
    post:
      consumes:
      - application/json
      description: Add a comment to the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Add wall comment
      tags:
      - user-profile
  /g/s/user-profile/{userId}/joined:
    get:
      consumes:
      - application/json
      description: Get list of users the target user is following
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Start offset
        in: query
        name: start
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user following list
      tags:
      - user-profile
  /g/s/user-profile/{userId}/member:
    delete:
      consumes:
      - application/json
      description: Unfollow the target user
      parameters:
      - description: User ID to unfollow
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Unfollow user
      tags:
      - user-profile
    get:
      consumes:
      - application/json
      description: Get list of users following the target user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Start offset
        in: query
        name: start
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user followers list
      tags:
      - user-profile
    post:
      consumes:
      - application/json
      description: Follow the target user
      parameters:
      - description: User ID to follow
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Follow user
      tags:
      - user-profile
...
  /x{comId}/s/user-profile/{userId}:
    get:
      consumes:
      - application/json
      description: Get public user profile information
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Community ID (NDC ID)
        in: path
        name: comId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user profile
      tags:
      - user-profile
    post:
      consumes:
      - application/json
      description: Update user profile details
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Community ID (NDC ID)
        in: path
        name: comId
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update user profile
      tags:
      - user-profile
...
  /x{comId}/s/user-profile/{userId}/comment:
    post:
      consumes:
      - application/json
      description: Reply to a comment on the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Community ID (NDC ID)
        in: path
        name: comId
        type: string
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Reply to wall comment
      tags:
      - user-profile
  /x{comId}/s/user-profile/{userId}/comment/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment from the user's wall
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Community ID (NDC ID)
        in: path
        name: comId
        type: string
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete wall comment
      tags:
      - user-profile
...
  /x{comId}/s/user-profile/{userId}/joined:
    get:
      consumes:
      - application/json
      description: Get list of users the target user is following
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Community ID (NDC ID)
        in: path
        name: comId
        type: string
      - description: Start offset
        in: query
        name: start
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get user following list
      tags:
      - user-profile
...
swagger: "2.0"